requirements.txt
fastapi~=0.116.0
uvicorn
sqlalchemy~=2.0.41
psycopg2-binary
python-dotenv~=1.1.1
pydantic~=2.11.7
langchain
langchain-openai~=0.3.27
pytest

langchain-core~=0.3.68